Revert "Merge branch 'feature/multitenant' into 'main'"
This reverts merge request !297

+++ app/Http/Controllers/Frontend/SedeController.php
@@ -6,129 +6,49 @@ use App\Http\Controllers\Controller;
 use App\Http\Controllers\Traits\CsvImportTrait;
 use App\Http\Requests\MassDestroySedeRequest;
 use App\Http\Requests\StoreSedeRequest;
+use App\Http\Requests\UpdateSedeRequest;
 use App\Models\Organizacion;
 use App\Models\Sede;
 use App\Models\Team;
 use Gate;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
-use Intervention\Image\Facades\Image;
 use Symfony\Component\HttpFoundation\Response;
-use Yajra\DataTables\Facades\DataTables;
 
 class SedeController extends Controller
 {
     use CsvImportTrait;
 
-    public function index(Request $request)
+    public function index()
     {
-        //abort_if(Gate::denies('configuracion_sede_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-        //dd( Sede::with(['organizacion', 'team'])->get());
-        if ($request->ajax()) {
-            $query = Sede::with(['organizacion', 'team'])->orderByDesc('id')->get();
-            $table = Datatables::of($query);
+        abort_if(Gate::denies('sede_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-            $table->addColumn('placeholder', '&nbsp;');
-            $table->addColumn('actions', '&nbsp;');
+        $sedes = Sede::all();
 
-            $table->editColumn('actions', function ($row) {
-                $viewGate = 'configuracion_sede_show';
-                $editGate = 'configuracion_sede_edit';
-                $deleteGate = 'configuracion_sede_delete';
-                $crudRoutePart = 'sedes';
+        $organizacions = Organizacion::get();
 
-                return view('partials.datatablesActions', compact(
-                    'viewGate',
-                    'editGate',
-                    'deleteGate',
-                    'crudRoutePart',
-                    'row'
-                ));
-            });
-
-            $table->editColumn('id', function ($row) {
-                return $row->id ? $row->id : '';
-            });
-            $table->editColumn('sede', function ($row) {
-                return $row->sede ? $row->sede : '';
-            });
-            $table->editColumn('foto_sedes', function ($row) {
-                return $row->foto_sedes ? $row->foto_sedes : '';
-            });
-            $table->editColumn('direccion', function ($row) {
-                return $row->direccion ? $row->direccion : '';
-            });
-            $table->editColumn('ubicacion', function ($row) {
-                //return "'lat' => ".$row->latitude. ",'long' => ".$row->longitud ? "'lat' => ".$row->latitude. ",'long' =>".$row->longitud : "";
-                return $row->id ? $row->id : '';
-            });
-            $table->editColumn('descripcion', function ($row) {
-                return $row->descripcion ? $row->descripcion : '';
-            });
-            $table->addColumn('organizacion_empresa', function ($row) {
-                return $row->organizacion ? $row->organizacion->empresa : '';
-            });
-
-            $table->rawColumns(['actions', 'placeholder', 'organizacion']);
-
-            return $table->make(true);
-        }
-
-        $organizacions = Organizacion::all();
-        //$org = $organizacions->organizacion;
-        //dd($organizacions->organizacion, $organizacions);
         $teams = Team::get();
-        $numero_sedes = Sede::count();
 
-        //$sede_inicio = !is_null($sedes) ? url('images/' . DB::table('organizacions')->select('logotipo')->first()->logotipo) : url('img/Silent4Business-Logo-Color.png');
-
-        return view('frontend.sedes.index', compact('organizacions', 'teams', 'numero_sedes'));
+        return view('frontend.sedes.index', compact('sedes', 'organizacions', 'teams'));
     }
 
     public function create()
     {
-        //abort_if(Gate::denies('configuracion_sede_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('sede_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $organizacions = Organizacion::get()->pluck('empresa', 'id')->prepend(trans('global.pleaseSelect'), '');
+        $organizacions = Organizacion::all()->pluck('empresa', 'id')->prepend(trans('global.pleaseSelect'), '');
 
         return view('frontend.sedes.create', compact('organizacions'));
     }
 
-    public function store(Request $request)
+    public function store(StoreSedeRequest $request)
     {
-
-        $client = new \GuzzleHttp\Client();
-        $geocoder = new \Spatie\Geocoder\Geocoder($client);
-        $geocoder->setApiKey('AIzaSyDByWqsyGQopJ8tnvFk8yp4PjcfG7zoXuo');
-        $result = $geocoder->getCoordinatesForAddress($request->direccion);
-        $request['latitude'] = $result['lat'];
-        $request['longitud'] = $result['lng'];
         $sede = Sede::create($request->all());
 
-        $image = null;
-        if ($request->file('foto_sedes') != null or !empty($request->file('foto_sedes'))) {
-            $extension = pathinfo($request->file('foto_sedes')->getClientOriginalName(), PATHINFO_EXTENSION);
-            $name_image = basename(pathinfo($request->file('foto_sedes')->getClientOriginalName(), PATHINFO_BASENAME), '.' . $extension);
-            $new_name_image = 'UID_' . $sede->id . '_' . $name_image . '.' . $extension;
-            $route = storage_path() . '/app/public/sedes/imagenes/' . $new_name_image;
-            $image = $new_name_image;
-            //Usamos image_intervention para disminuir el peso de la imagen
-            $img_intervention = Image::make($request->file('foto_sedes'));
-            $img_intervention->resize(256, null, function ($constraint) {
-                $constraint->aspectRatio();
-            })->save($route);
-        }
-
-        $sede->update([
-            'foto_sedes' => $image,
-        ]);
-
-        return redirect()->route('frontend.sedes.index')->with('success', 'Guardado con éxito');
+        return redirect()->route('frontend.sedes.index');
     }
 
     public function edit(Sede $sede)
     {
-        //abort_if(Gate::denies('configuracion_sede_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('sede_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $organizacions = Organizacion::all()->pluck('empresa', 'id')->prepend(trans('global.pleaseSelect'), '');
 
@@ -137,49 +57,16 @@ class SedeController extends Controller
         return view('frontend.sedes.edit', compact('organizacions', 'sede'));
     }
 
-    public function update(Request $request, $id)
+    public function update(UpdateSedeRequest $request, Sede $sede)
     {
-        // $sede->update($request->all());
+        $sede->update($request->all());
 
-        $sede = Sede::find($id);
-        $image = $sede->foto_sedes;
-        if ($request->file('foto_sedes') != null or !empty($request->file('foto_sedes'))) {
-
-            //Si existe la imagen entonces se elimina al editarla
-
-            $isExists = Storage::disk('public')->exists('sedes/imagenes/' . $sede->foto_sedes);
-            if ($isExists) {
-                if ($sede->foto_sedes != null) {
-                    unlink(storage_path('/app/public/sedes/imagenes/' . $sede->foto_sedes));
-                }
-            }
-            $extension = pathinfo($request->file('foto_sedes')->getClientOriginalName(), PATHINFO_EXTENSION);
-            $name_image = basename(pathinfo($request->file('foto_sedes')->getClientOriginalName(), PATHINFO_BASENAME), '.' . $extension);
-            $new_name_image = 'UID_' . $sede->id . '_' . $name_image . '.' . $extension;
-            $route = storage_path() . '/app/public/sedes/imagenes/' . $new_name_image;
-            $image = $new_name_image;
-            //Usamos image_intervention para disminuir el peso de la imagen
-            $img_intervention = Image::make($request->file('foto_sedes'));
-            $img_intervention->resize(256, null, function ($constraint) {
-                $constraint->aspectRatio();
-            })->save($route);
-        }
-
-        $sede->update([
-
-            'sede' =>  $request->sede,
-            'foto_sedes' =>  $request->foto_sede,
-            'direccion' =>  $request->direccion,
-            'descripcion' =>  $request->descripcion,
-            'foto_sedes' => $image,
-        ]);
-
-        return redirect()->route('frontend.sedes.index')->with('success', 'Editado con éxito');
+        return redirect()->route('frontend.sedes.index');
     }
 
     public function show(Sede $sede)
     {
-        //abort_if(Gate::denies('configuracion_sede_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('sede_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $sede->load('organizacion', 'team');
 
@@ -188,11 +75,11 @@ class SedeController extends Controller
 
     public function destroy(Sede $sede)
     {
-        //abort_if(Gate::denies('configuracion_sede_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('sede_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $sede->delete();
 
-        return back()->with('deleted', 'Registro eliminado con éxito');
+        return back();
     }
 
     public function massDestroy(MassDestroySedeRequest $request)
@@ -200,33 +87,5 @@ class SedeController extends Controller
         Sede::whereIn('id', request('ids'))->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
-    }
-
-    public function obtenerListaSedes(Sede $sedes)
-    {
-        //abort_if(Gate::denies('organizacion_sede_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-        //$sede = Sede::get();
-        $sede = Sede::paginate(3);
-        $organizacions = Organizacion::get();
-        $teams = Team::get();
-        $numero_sedes = Sede::count();
-
-        return view('frontend.sedes.sedes-organizacion', compact('sede', 'organizacions', 'teams', 'numero_sedes'));
-    }
-
-    public function ubicacion($request)
-    {
-        //abort_if(Gate::denies('organizacion_sede_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-        $sede = Sede::find($request);
-
-        return view('frontend.sedes.ubicacion', compact('sede'));
-    }
-
-    public function ubicacionorg($request)
-    {
-        //abort_if(Gate::denies('organizacion_sede_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-        $sede = Sede::find($request);
-        //dd($sede);
-        return view('frontend.sedes.ubicacion', compact('sede'));
     }
 }
